remove similar manga background job
