+++ app/src/main/res/layout/manga_header_item.xml
@@ -270,7 +270,6 @@
                 android:layout_marginEnd="6dp"
                 android:padding="5dp"
                 android:tooltipText="@string/similar"
-                android:visibility="gone"
                 app:iiv_icon="gmf_account_tree"
                 tools:background="@color/md_black_1000"
                 tools:visibility="visible" />
